// Frontend/src/app/shared/interfaces/fornecedor.ts
export interface Fornecedor {
    codigoEscritorio: number;
    codigoTipoFornecedor: number;
    codigoProfissional: number;
    codigoBanco: number;
    id: number;
    nomeFornecedor: string;
    codigoFornecedorSap: string;
    nomeTipoFornecedor: string;
    nomeEscritorio: string;
    nomeProfissional: string;
    nomeBanco: string;
    selected?: boolean;
}